requirements.txt
argcomplete==1.10.0
azure-core==1.9.0
azure-search-documents==11.1.0b4
beautifulsoup4==4.8.0
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
chardet==3.0.4
cymem==2.0.5
docx2txt==0.8
EbookLib==0.17.1
en-core-web-sm==2.3.1
extract-msg==0.23.1
fasttext==0.9.2
idna==2.10
IMAPClient==2.1.0
isodate==0.6.0
lxml==4.6.2
msrest==0.6.19
murmurhash==1.0.5
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
pandas==1.2.0
pdfminer.six==20181108
Pillow==8.1.0
plac==1.1.3
preshed==3.0.5
pybind11==2.6.1
pycryptodome==3.9.9
python-dateutil==2.8.1
python-pptx==0.6.18
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.12.0
sortedcontainers==2.3.0
soupsieve==2.1
spacy==2.3.5
SpeechRecognition==3.8.1
srsly==1.0.5
textract==1.6.3
thinc==7.4.5
tqdm==4.55.2
typing-extensions==3.7.4.3
tzlocal==1.5.1
urllib3==1.26.2
wasabi==0.8.0
xlrd==1.2.0
XlsxWriter==1.3.7
